+++ client/src/Product/pages/ProductList/index.tsx
@@ -1,10 +1,8 @@
 import React, { useEffect, useState, Dispatch, SetStateAction } from 'react'
 import axios, { AxiosResponse } from 'axios'
-import { Table } from 'antd'
 import { useRouteMatch, useHistory } from 'react-router-dom'
-import * as H from 'history';
 import PageTitle from 'src/shared/components/PageTitle'
-import TableHeader from '../../components/TableHeader'
+import ProductTable from '../../components/ProductTable'
 
 export interface IProduct {
   id: number
@@ -13,67 +11,10 @@ export interface IProduct {
   price: number
 }
 
-const buildProductColumns = (products: IProduct[], setProducts: Dispatch<SetStateAction<IProduct[]>>, history: H.History<H.LocationState>) => {
-  const deleteProduct = async (id: number) => {
-    await axios.delete(`http://localhost:3000/api/v1/products/${id}`)
-    const nextProductList: IProduct[] = products.filter((product: IProduct) => ( product.id !== id ))
-    setProducts(nextProductList)
-    history.push('/products')
-  }
-
-  return (
-    [
-      {
-        title: '名前',
-        dataIndex: 'name',
-        key: 'name',
-      },
-      {
-        title: 'メーカー',
-        dataIndex: 'maker',
-        key: 'maker',
-      },
-      {
-        title: '価格',
-        dataIndex: 'price',
-        key: 'price',
-      },
-      {
-        title: '',
-        key: 'action',
-        render: (record: any) => {
-          return (
-            <button onClick={() => {
-              history.push(`/products/${record.key}`)
-            }}>
-              編集
-            </button>
-          )
-        },
-      },
-      {
-        title: '',
-        key: 'action',
-        render: (record: any) => {
-          return (
-            <button onClick={() => {
-              deleteProduct(record.key) 
-            }}>
-              削除
-            </button>
-          )
-        },
-      },
-    ]
-  )
-}
-
 const ProductList: React.SFC = () => {
   const [products, setProducts] : any = useState([])
   let match = useRouteMatch()
   let history = useHistory()
-
-  const onCreateClick = () => history.push(`${match.url}/new`)
 
   useEffect(() => {
     const getApiResult = async () => {
@@ -83,20 +24,13 @@ const ProductList: React.SFC = () => {
     getApiResult()
   }, [])
 
-  const dataSource = products.map((product: IProduct) => (
-    {
-      key: product.id,
-      name: product.name,
-      maker: product.maker,
-      price: product.price,
-    }
-  ))
-
   return (
     <>
       <PageTitle>商品一覧</PageTitle>
-      <TableHeader onCreateClick={onCreateClick} />
-      <Table dataSource={dataSource} columns={buildProductColumns(products, setProducts, history)}/>
+      <ProductTable
+        products={products}
+        setProducts={setProducts}
+      />
     </>
   )
 }
